// example/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as admin from "../admin.js";
import type * as debug from "../debug.js";
import type * as example from "../example.js";
import type * as http from "../http.js";
import type * as nestedWorkflow from "../nestedWorkflow.js";
import type * as passingSignals from "../passingSignals.js";
import type * as transcription from "../transcription.js";
import type * as userConfirmation from "../userConfirmation.js";
import type * as webhookExample from "../webhookExample.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  debug: typeof debug;
  example: typeof example;
  http: typeof http;
  nestedWorkflow: typeof nestedWorkflow;
  passingSignals: typeof passingSignals;
  transcription: typeof transcription;
  userConfirmation: typeof userConfirmation;
  webhookExample: typeof webhookExample;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {
  workflow: import("@convex-dev/workflow/_generated/component.js").ComponentApi<"workflow">;
};